New option: shuffle world order.
This will keep levels within each world the same, but shuffle the order
in which the worlds appear. So for example, the first world you play
could be world 4 in the vanilla game.
World 8 will always be last.
The 'Pipe Transitions' option works as expected.
I also improved the way I store world/level data in the randomizer.