Added custom code to allow warp zones anywhere!
Long story short: I wrote custom 6502 code that replaces the original
code that's run when you enter a warp zone. The new code looks at a
different value in memory to determine which warp zone you are in, and
no longer relies on which world you are in. Therefore, this allows 1-2
and 4-2 to be placed anywhere in the world layout!!!!

I might make a separate option in the future to let the player choose
whether to allow this or use the old way where 1-2 has to be in world 1
and 4-2 in world >= 2.

I haven't looked at how many cpu cycles the new code uses vs the old
code, but does it matter? The difference is minimal at best, maybe zero.

The custom code is 15 bytes (including 1 NOP, so technically 14 bytes).
The original code was also 15 bytes, therefore the NOP.