Make sure the last byte of the ROM is included when you save.
Due to an inaccuracy when base64-decoding the ROM data, the very last
byte of the ROM file wasn't saved. This has no real impact on the game
itself, but could make the ROM be impossible to load in emulators that
check the file size against the header information and refuses to load
the ROM if there's a mismatch.

Fixed this simply by pushing that byte to the end of the array.